requirements.txt
numpy
dbus-python
PyGObject
cryptography
requests==2.31.0
bmp280==0.0.4
pycryptodome>=3.10.1